+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 `lineman` uses [PVP Versioning][1].
 The changelog is available [on GitHub][2].
 
+## 1.0.0
+
+* Remove stack configuration
+* Land config on `dhall`
 
 ## 0.0.7